Adapt solution to previous commit
[Refactor] usage of int everywhere to use ulong
[Create] TextualTokenExtensions to fill in for old system
[Refactor] usage of Lexer and LexerToken to new types everywhere

+++ BakedEnv/Interpreter/Instructions/ControlStatementInstruction.cs
@@ -10,7 +10,7 @@ public class ControlStatementInstruction : InterpreterInstruction
     public BakedExpression[] Parameters { get; }
     public IEnumerable<InterpreterInstruction> Instructions { get; }
     
-    public ControlStatementInstruction(ControlStatementDefinition definition, BakedExpression[] parameters, IEnumerable<InterpreterInstruction> instructions, int sourceIndex) : base(sourceIndex)
+    public ControlStatementInstruction(ControlStatementDefinition definition, BakedExpression[] parameters, IEnumerable<InterpreterInstruction> instructions, ulong sourceIndex) : base(sourceIndex)
     {
         StatementDefinition = definition;
         Parameters = parameters;
