+++ BakedEnv/Interpreter/IntermediateTokens/IntermediateToken.cs
@@ -4,7 +4,7 @@ public abstract class IntermediateToken
 {
     public bool IsComplete { get; set; }
 
-    public abstract int StartIndex { get; }
+    public abstract ulong StartIndex { get; }
     public abstract int Length { get; }
-    public abstract int EndIndex { get; }
+    public abstract ulong EndIndex { get; }
 }
